+++ example/integration_test/_test_example_app.dart
@@ -79,6 +79,37 @@ Future<void> integrationTesting(WidgetTester tester) async {
   /// Simulate some events (eg. paused and resumed the app)
   await testEventHandling(tester);
 
+  /// **IMPORTANT** If a app returns from another app the current State is destroyed
+  /// i.e. After a resumed event the current State object is unmounted
+  /// A new State object is created!
+  /// And so that state variable above has a useless State object now!
+
+  /// Test that a state object as been replaced!
+  expect(state.mounted, isFalse, reason: _location);
+
+  expect(state.deactivated, isTrue, reason: _location);
+
+  expect(state.inactive, isTrue, reason: _location);
+
+  expect(state.detached, isTrue, reason: _location);
+
+  // Even the listener will be gone if the State is now disposed.
+  final contains = state.afterContains(listener);
+
+  // The system will dispose of the State at its own discretion
+  // You'll have no idea if and when that is. Don't test for disposed!
+
+//  expect(contains, isFalse, reason: _location);
+
+//  expect(state.disposed, isTrue, reason: _location);
+
+  /// A new State object has been introduced!
+  state = con.state!;
+
+  expect(state.mounted, isTrue, reason: _location);
+
+  expect(state, isA<Page1State>(), reason: _location);
+
   /// Go to Page 2
   await tester.tap(find.byKey(const Key('Page 2')));
   await tester.pumpAndSettle(const Duration(milliseconds: 200));